requirements.txt
django==4.2.16
django-extensions==3.2.3
networkx==3.1
telethon==1.37.0
python-decouple==3.8
django-colorfield==0.11.0

beautifulsoup4==4.10.0
lxml==4.8.0
scipy==1.8.0
numpy==1.22.3
